test: move libinfo stale-request fallback tests into dns-prefetch.test.ts
